Ability to choose disciplines for concrete student
Added a list of students at the Home page. The "Select" page where all available disciplines and their status shown is opened by clicking on the "Disciplines" button for concrete student . Each discipline has a checkbox which show if it choosed or not. To save changes the "OK" button must be clicked.